feat(search): moved spinner inside search button + improved layout
